Added Scoreboard and option to restart game or quit
The player is shown an ending/quit/goodbye message when the game ends.

Ask the user if they would like to play again.

If yes, then cycle the user back into guessing a number again.
If no, then provide the user with an ending/quit/goodbye message.

The player is shown their number of tries/their score at the end of each game once they have guessed the correct number.